Added error handling for no value path

+++ dask/dataframe/io/parquet/core.py
@@ -477,7 +477,9 @@ def to_parquet(
     # Trim any protocol information from the path before forwarding
     path = fs._strip_protocol(path)
 
-    if overwrite:
+    if overwrite and path:
+        if path in ["", ".", "./", "/"]:
+            raise ValueError("If using overwrite=True, path must be an explicit directory!!")
         fs.rm(path, recursive=True)
 
     # Save divisions and corresponding index name. This is necessary,
